Moves permission component to audio. Updates imports. Puts feature flag on app screen.
